opsgenie: skip importing escalation policy whose team isn't known (#64)
When performing selective import of teams, we still query the escalation
policies globally (in case there are escalation policies which aren't
related to any team wants to be imported).

However, when an escalation policy _is_ linked to a team that isn't
selected to be imported, we should just skip it because we know we don't
want them. Prior this changeset, the migrator would hard fail and not
allow users to continue. With this changeset, the migrator will ignore
the entry and continue the progress.